jkh w_app.js ip checking system make
