+++ web_server/api/router.js
@@ -15,18 +15,8 @@ router.get('/login' ,(req, res) => {
 
 
 //'//web/landing/industry/index.html' 일때 로그인의 유무를 판단하는 기능 구현
-router.get('/', (req, res) => {    
-    var rrq_ip = jkh_f.ip_denying(req);
-    if (rrq_ip.state == 1) {
-        iplist.push(rrq_ip.ip);
-        console.log(`ip 차단 : ${rrq_ip.ip}`);
-        jkh_f.webhook('warn', `${req.ip} web '/' enter and denying`);
-    }
-    const str = 'web server gate';
-    jkh_f.webhook('success', `${req.ip} web '/' enter`);
-    res.redirect(302, '/w/user/index.html');
-    return res.send(str);    
-});
+//router
+// '/' get 요청 작성 금지
 //메인페이지로 이동
 
 /*****************************/
